[BEAM-5435] Kryo separated from euphoria-core to extensions-kryo.
