+++ sdks/java/extensions/euphoria/euphoria-core/src/test/java/org/apache/beam/sdk/extensions/euphoria/core/client/operator/OperatorTests.java
@@ -22,10 +22,10 @@ import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.extensions.euphoria.core.client.dataset.Dataset;
 import org.apache.beam.sdk.extensions.euphoria.core.client.operator.base.Operator;
 import org.apache.beam.sdk.extensions.euphoria.core.client.type.TypeAwares;
-import org.apache.beam.sdk.extensions.euphoria.core.coder.KryoCoder;
 import org.apache.beam.sdk.extensions.euphoria.core.translate.EuphoriaOptions;
 import org.apache.beam.sdk.extensions.euphoria.core.translate.OperatorTranslator;
 import org.apache.beam.sdk.extensions.euphoria.core.translate.TranslatorProvider;
+import org.apache.beam.sdk.extensions.kryo.KryoCoder;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.testing.TestPipeline;
