+++ sdks/java/extensions/kryo/build.gradle
@@ -16,5 +16,24 @@
  * limitations under the License.
  */
 
-/** Euphoria API focused on Beam coders. */
-package org.apache.beam.sdk.extensions.euphoria.core.coder;
+apply plugin: org.apache.beam.gradle.BeamModulePlugin
+applyJavaNature()
+
+description = "Apache Beam :: SDKs :: Java :: Extensions :: Euphoria Java 8 DSL :: Kryo"
+
+ext {
+    kryoVersion = '4.0.2'
+}
+
+dependencies {
+    shadow project(path: ":beam-sdks-java-core", configuration: "shadow")
+    shadow "com.esotericsoftware:kryo:${kryoVersion}"
+    testCompile project(path: ":beam-sdks-java-core", configuration: "shadowTest")
+    testCompile project(':beam-runners-direct-java')
+}
+
+test {
+    jvmArgs '-Dsun.io.serialization.extendedDebugInfo=true'
+}
+
+test.testLogging.showStandardStreams = false
