+++ sdks/java/extensions/kryo/src/main/java/org/apache/beam/sdk/extensions/kryo/IdentifiedRegistrar.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.extensions.euphoria.core.coder;
+package org.apache.beam.sdk.extensions.kryo;
 
 import com.esotericsoftware.kryo.Kryo;
 import java.io.Serializable;
