+++ sdks/java/extensions/kryo/src/test/java/org/apache/beam/sdk/extensions/kryo/RegisterCodersTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.extensions.euphoria.core.coder;
+package org.apache.beam.sdk.extensions.kryo;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
