+++ settings.gradle
@@ -82,7 +82,8 @@ include "beam-sdks-java-extensions-euphoria-core"
 project(":beam-sdks-java-extensions-euphoria-core").dir = file("sdks/java/extensions/euphoria/euphoria-core")
 include "beam-sdks-java-extensions-euphoria-testing"
 project(":beam-sdks-java-extensions-euphoria-testing").dir = file("sdks/java/extensions/euphoria/euphoria-testing")
-
+include "beam-sdks-java-extensions-kryo"
+project(":beam-sdks-java-extensions-kryo").dir = file("sdks/java/extensions/kryo")
 include "beam-sdks-java-extensions-google-cloud-platform-core"
 project(":beam-sdks-java-extensions-google-cloud-platform-core").dir = file("sdks/java/extensions/google-cloud-platform-core")
 include "beam-sdks-java-extensions-json-jackson"
